requirements.txt
pygame==2.6.0
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9

numpy==2.2.6